+++ src/fitModels.hpp
@@ -1004,7 +1004,7 @@ class lm_output
 		
 		std::vector<double> beta;
 		std::vector<double> se;
-		std::vector<double> pval;
+		std::vector<long double> pval;
     std::vector<double> log_pval;
 		std::vector<double> vif;
 		
@@ -1012,7 +1012,7 @@ class lm_output
 		std::string name;
 		std::string info;
 		
-		void push_back(const double& b, const double& s, const double& p, const double& log_p)
+		void push_back(const double& b, const double& s, const long double& p, const double& log_p)
 		{
 			beta.push_back(b);
 			se.push_back(s);
@@ -1020,7 +1020,7 @@ class lm_output
       log_pval.push_back(log_p);
 		};
 
-		void push_back(const double& b, const double& s, const double& p, const double& log_p, const double& v)
+		void push_back(const double& b, const double& s, const long double& p, const double& log_p, const double& v)
 		{
 			beta.push_back(b);
 			se.push_back(s);
@@ -1057,10 +1057,10 @@ class forward_lm
 		std::vector<double> se;
 		std::vector<double> beta_0;
 		std::vector<double> se_0;
-		std::vector<double> pval_0;
-		std::vector<double> pval_seq;
-		std::vector<double> pval_joint;
-		std::vector<double> pval_adj;
+		std::vector<long double> pval_0;
+		std::vector<long double> pval_seq;
+		std::vector<long double> pval_joint;
+		std::vector<long double> pval_adj;
     std::vector<double> log_pval_0;
     std::vector<double> log_pval_seq;
     std::vector<double> log_pval_joint;
